+++ Sources/Core.swift
@@ -203,7 +203,7 @@ typealias Error = ErrorType
 
 /// Errors, that may occur while resolving from a `Provider`.
 /// - ToDo: Implement `case cyclicDependency`
-public enum InjectionError<Key: ProvidableKey>: Error {
+public enum InjectionError<Key: ProvidableKey>: Error, Equatable {
     #if swift(>=3.0)
     #else
     public typealias Error = ErrorType
@@ -223,4 +223,24 @@ public enum InjectionError<Key: ProvidableKey>: Error {
     /// The internal `InjectedProvider` has the wrong type.
     /// This is likely a bug in the framework.
     case invalidInjection(key: Key, injected: Any, expected: Any.Type)
+}
+
+/// Tests for equality.
+/// Ignores: 
+///     - `InjectionError.nonMatchingType(provided:expected:)`'s expected
+///     - `InjectionError.customError(_)`'s parameter
+///     - `InjectionError.invalidInjection(key:injected:expected:)`'s injected
+public func ==<K: ProvidableKey>(lhs: InjectionError<K>, rhs: InjectionError<K>) -> Bool {
+    switch (lhs, rhs) {
+    case let (.keyNotProvided(lk), .keyNotProvided(rk)):
+        return lk == rk
+    case let (.nonMatchingType(_, le), .nonMatchingType(_, re)):
+        return le == re
+    case (.customError(_), .customError(_)):
+        return true
+    case let (.invalidInjection(lk, _, le), .invalidInjection(rk, _, re)):
+        return lk == rk && le == re
+    default:
+        return false
+    }
 }